moved shared logic in python api client
