+++ packages/python-sdk/e2b/sandbox_sync/sandbox_api.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 from httpx import HTTPTransport
 from packaging.version import Version
 
-from e2b.sandbox.sandbox_api import SandboxInfo, SandboxApiBase
+from e2b.sandbox.sandbox_api import ListSandboxesResponse, SandboxInfo, SandboxApiBase
 from e2b.exceptions import TemplateException
 from e2b.api import ApiClient, SandboxCreateResponse, handle_api_exception
 from e2b.api.client.models import NewSandbox, PostSandboxesSandboxIDTimeoutBody
@@ -29,13 +29,6 @@ from e2b.exceptions import NotFoundException, TemplateException
 from e2b.sandbox.sandbox_api import SandboxApiBase, SandboxInfo, SandboxMetrics
 from httpx import HTTPTransport
 from packaging.version import Version
-
-class ListSandboxesResponse:
-     def __init__(self, sandboxes: List[SandboxInfo], has_more_items: bool, next_token: Optional[str], iterator: Generator[SandboxInfo, None, None]):
-         self.sandboxes = sandboxes
-         self.has_more_items = has_more_items
-         self.next_token = next_token
-         self.iterator = iterator
 
 
 class SandboxApi(SandboxApiBase):
@@ -116,19 +109,7 @@ class SandboxApi(SandboxApiBase):
             has_more_items = bool(token)
 
             sandboxes = [
-                SandboxInfo(
-                    sandbox_id=SandboxApi._get_sandbox_id(
-                        sandbox.sandbox_id,
-                        sandbox.client_id,
-                    ),
-                    template_id=sandbox.template_id,
-                    name=sandbox.alias if isinstance(sandbox.alias, str) else None,
-                    metadata=(
-                        sandbox.metadata if isinstance(sandbox.metadata, dict) else {}
-                    ),
-                    started_at=sandbox.started_at,
-                    state=sandbox.state,
-                )
+                SandboxInfo.from_listed_sandbox(sandbox)
                 for sandbox in res.parsed
             ]
